refactor: extract UrgeMath from PersonaDecisionService
Moves the four pure pre-gate helpers (CalculateResponseUrge,
CountRecentSelfMessages, CalculateNetPressure, IsObviousSkip) onto a new
static UrgeMath class in PartyTown.Services.ResponsePipeline. PersonaGrain's
pre-gate now calls UrgeMath.* directly instead of pulling in the LLM-call
service for unrelated reasons; PersonaDecisionService keeps the
ShouldRespondAsync orchestration and the ResponseUrge record.

Migrates the math-only test cases from PersonaDecisionServiceTest into a
new UrgeMathTest. The end-to-end ShouldRespondAsync coverage stays on
PersonaDecisionServiceTest.

No behavior change.

Closes #60
